perf: migrate inverted-index posting lists to RoaringBitmap
Follow-up to the sorted-merge engine. Store's 11 inverted indexes change
value type from Vec<u32> to roaring::RoaringBitmap. Posting lists are now
compressed bitmaps: dense/high-cardinality lists (e.g. every record under
one eventName) shrink substantially, and the query engine gets native
SIMD AND/OR/NOT instead of hand-rolled sorted-merge.

engine.rs: compute_and_group/compute_matching_ids now use roaring's `&`,
`|`, `-` operators and insert_range for the negative-only base set; the
custom intersect/union/difference helpers are gone. Bitmap iteration is
ascending-by-id, same order the old Vec produced, so result ordering and
detection-rule behavior are unchanged.

The index_for accessor added in the previous commit contained the field->
index churn: only the storage type + the ~60 read-sites that iterate a
posting list changed (for &id -> for id; .len() -> .len() as usize where a
usize is needed; matching_record_ids: <bitmap> -> .iter().collect();
extend_from_slice -> extend; build sites .push -> .insert). Vec<u32> query
results and session-index lists are untouched.

All 142 core tests pass; workspace compiles with zero warnings.

+++ crates/app/src/commands/geoip.rs
@@ -193,7 +193,7 @@ pub async fn list_ips(
         } else {
             store.idx_source_ip
                 .iter()
-                .map(|(ip, ids)| (ip.to_string(), ids.len()))
+                .map(|(ip, ids)| (ip.to_string(), ids.len() as usize))
                 .collect::<std::collections::HashMap<String, usize>>()
         }
     };
